Refactor scripts to use real temp directory
This set of changes completely overhauls _where_ stuff is done from hard
coded paths in /tmp to a real temporary directory. The build_all script
will do it's own garbage collection after running. Anything that gets
explicitly copied to a target location (using `install` so permissions
and potentially ownership can be set right the first time rather than
later) will be retained, everything else is discarded. The exception is
a zip file with some of the key pieces from the compile cycle and a
report will be wrapped up and dropped it $output (or `pwd` if not set).

In the mean time the various target locations such as the output
directory, a directory to stash dated draft versions, where the api is
hosted, etc. have all been made configurable from the run time
environment.

Additionally rather than manually configuring the TeX environment on
every run, the assumption is now that there is a working ConTeXt in your
path. This will break the current install (at least on test.door43.org)
as a version has been manually put in /opt with some other parts of the
install copied local to the repository). The proper fix is to do a
system wide installation of texlive from the normal distribution
channels. Baring a system wide install that just works, a script should
be included in the repository that gets a local one setup properly.