Adjust scripts to use TeX includes from local repo
There used to be a Door43/includes repo that had some TeX snippets
relating to language. These may prove useful for projects other than
OBS, but for now it seems they really only work with this data set. Up
until now the build scripts assumed that a symlink was available to this
repo from the root of this one, but with the repo only existing on the
test server that wasn't a very good assumption.

The previous commits are all rebased cherry-picks from that rep to pull
them in local to this one. The result is that the TeX related templating
stuff is all in the same place as the scripts that use it.

In the future these could be split out to a set of TeX classes that
could be separate and re-used by other projects, but as they are in no
shape for that yet and as these scripts need an exact match to the
versions expected, having them local to this repo seems to make the most
sense.