Parameterize 'checking level'
Previously the checking level was hard coded as '3' in the TeX template.
This adds a new parameter that can be set one of two ways:

1. The JSON meta data file can set it as a key 'checkinglevel'. Note
   none of the current data files are passing this data, but they now
   could and any generated PDF's would show the correct level
   automatically.

2. When running the export.py script to generate the TeX document a
   command line option is available `-c` or `--checkinglevel`. Setting
   this to either 1, 2 or 3 will set a value in the event the JSON data
   doesn't specify.

The default value has been set to 1 so that documents don't get marked
as higher than they deserve. If a document is ready to be marked higher
the API feeding us the data should say so!