Add _hack_ for localized TOC title
The JSON meta data does not appear to have any localized string that
corresponds to a title for the index page. Up until now this has been
hard coded in English in the ConTeXt template.

This is not a fix! This is a hack! But it notes the issue is such a way
as to make the output less egregious while waiting for proper API support.

This uses a Regular Expression to try to rip out an appropriate title
from the localized front-matter string. If it finds one great. If it
doesn't find an exact match it will pass in the English. However none of
this will ever run _IN THE EVENT_ the API actually returns a localized
string for this use (key of 'toctitle').