Fix broken Regex
Unicode codepoint U+002D is the normal hypen-minus you find on your ASCI
keyboard. In the context of Regex character classes this is used to
denote a range unless it is escaped or the first character in a set. Not
only was it included twice in this set (once as plain '-' and once as a
unicode escape sequence) but the latter was causing it to match a range
of characters (U+2013 to U+FE63) which includes a ton of punctuation
marks, including the single apostrophe used in some languages inside
book names (e.g. Turkish). This was causing all sorts of assorted bits
to get replaced with a figure-dash:

> Öykünün Kutsal Kitap‒taki yeri: Matta 24:14

(That is a ‘ mark in the source.)