Split front matter and insert checking level
The PDFs produced so far have been manually tweaked to move the checking
level into the middle of what in otherwise the front-matter string. This
attempts to do this automatically, but should fail gracefully if the
Regex doesn't work in a language.

However, THIS IS A HACK! The real fix here is that the front-matter
string needs to be broken up and served up in pieces so it can be more
appropriately positioned for various output formats.