fix margin override
The pandoc template had a manual override for margins that was a)
logically prior to and b) conflicting with what pandoc expects to do for
that particular case.

This goes back to using the pandoc method. If you did want to hard code
it in the template you would want to take out the pandoc special
replacements for loading the geometry package.